refactor: move Tasklist index descriptors outside tasklist package
