+++ tasklist/importer/src/main/java/io/camunda/tasklist/zeebeimport/ImportPositionHolderAbstract.java
@@ -10,7 +10,7 @@ package io.camunda.tasklist.zeebeimport;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.tasklist.Metrics;
 import io.camunda.tasklist.property.TasklistProperties;
-import io.camunda.webapps.schema.descriptors.tasklist.index.TasklistImportPositionIndex;
+import io.camunda.webapps.schema.descriptors.index.TasklistImportPositionIndex;
 import io.camunda.webapps.schema.entities.ImportPositionEntity;
 import jakarta.annotation.PostConstruct;
 import java.io.IOException;
