+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/TasklistMetricIndex.java
@@ -5,36 +5,28 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.tasklist.index;
+package io.camunda.webapps.schema.descriptors.index;
 
+import static io.camunda.webapps.schema.descriptors.ComponentNames.TASK_LIST;
+
+import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
 import io.camunda.webapps.schema.descriptors.IndexDescriptor;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
-import io.camunda.webapps.schema.descriptors.tasklist.TasklistIndexDescriptor;
 import java.util.Optional;
 
-public class FormIndex extends TasklistIndexDescriptor implements Prio5Backup {
+public class TasklistMetricIndex extends AbstractIndexDescriptor implements Prio5Backup {
 
-  public static final String INDEX_NAME = "form";
-  public static final String INDEX_VERSION = "8.4.0";
+  public static final String INDEX_NAME = "metric";
+  public static final String INDEX_VERSION = "8.3.0";
 
   public static final String ID = "id";
-  public static final String KEY = "key";
-  public static final String BPMN_ID = "bpmnId";
-  public static final String SCHEMA = "schema";
+  public static final String EVENT = "event";
+  public static final String VALUE = "value";
+  public static final String EVENT_TIME = "eventTime";
   public static final String TENANT_ID = IndexDescriptor.TENANT_ID;
-  public static final String VERSION = "version";
-  public static final String IS_DELETED = "isDeleted";
 
-  public static final String PROCESS_DEFINITION_ID = "processDefinitionId";
-  public static final String EMBEDDED = "embedded";
-
-  public FormIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public TasklistMetricIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
-  }
-
-  @Override
-  public String getVersion() {
-    return INDEX_VERSION;
   }
 
   @Override
@@ -45,5 +37,15 @@ public class FormIndex extends TasklistIndexDescriptor implements Prio5Backup {
   @Override
   public Optional<String> getTenantIdField() {
     return Optional.of(TENANT_ID);
+  }
+
+  @Override
+  public String getVersion() {
+    return INDEX_VERSION;
+  }
+
+  @Override
+  public String getComponentName() {
+    return TASK_LIST.toString();
   }
 }
