+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/utils/SearchDBExtension.java
@@ -8,8 +8,8 @@
 package io.camunda.exporter.utils;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
 import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.extension.AfterEachCallback;
